feat: make new errors also accessible to python
